Milestone and Task Information with Charts
Contains three sheets. The first is labeled all and contains all of the rough information about every task and milestone. The second is milestone. Contains all the information, including descriptions from report, about each milestone and a Gantt chart of the milestones. The third is tasks and contains all the information, including descriptions from report, about each task, with tasks grouped by the milestone that they most predominantly fall under, and a Gantt chart of the tasks broken down by which milestone it falls under..